refactor: debounce.jsx to typescript and moving it to Hooks
